fix color division (see 3.2.19/3.3,14/3.4.0rc3)

+++ src/Compiler.php
@@ -1362,7 +1362,7 @@ class Compiler
                     if ($rval == 0) {
                         $this->throwError("color: Can't divide by zero");
                     }
-                    $out[] = $lval / $rval;
+                    $out[] = (int) ($lval / $rval);
                     break;
                 case '==':
                     return $this->opEq($left, $right);
